\name{breast}
\alias{breast}
\docType{data}

\title{Example dataset for learning gene regulatory network.} 

\description{
Breast cancer is one of the most prevalent types of cancer which can be classified into four molecular subtypes, namely, luminal A, basal-like, HER2-enriched, and lu- minal B, based on their tumor expression profiles (Haque et al, 2012). In this study, we aim to construct a single gene regulatory network across the four subtypes to discover the overall gene regulation mechanism in breast cancer. It should be used in \code{BRGM(data,...)}.
}
\usage{data(breast)}
\format{
\describe{
  \item{ breast }{ a \emph{n}x\emph{p} breast cancer gene expression data.}
}}


\references{
Haque, R., Ahmed, S. A., Inzhakova, G., Shi, J., Avila, C., Polikoff, J., ... and Press, M. F.
(2012). Impact of breast cancer subtypes and treatment on survival: an analysis spanning two decades. Cancer Epidemiology and Prevention Biomarkers, 21(10), 1848-1855.
}


\keyword{datasets}
